pyproject.toml
[project]
name = "fhirside-chat"
version = "0.1.0"
description = "Chat interface to a Fhir server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai-slim[openai]>=1.0.15",
    "openai>=1.6.0",
    "fastmcp>=2.0.0",
    "rich>=13.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.37.0",
    "sqlmodel>=0.0.27",
]

[build-system]
requires = ["setuptools>=42.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[dependency-groups]
dev = [
    "ipython>=9.6.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.3",
]
